requirements.txt
alembic==0.9.2
amqp==2.1.4
aniso8601==1.3.0
apns==2.0.1
apturl==0.5.2
asn1crypto==0.22.0
Babel==2.6.0
backports-abc==0.5
backports.ssl-match-hostname==3.5.0.1
billiard==3.5.0.4
Brlapi==0.6.6
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
defusedxml==0.5.0
distro-info==0.18
Django==2.1.4
django-allauth==0.38.0
django-appconf==1.0.2
django-node==4.0.0
django-webpack==3.1.1
django-webpack-helper==0.2.2
django-webpack-loader==0.6.0
django-webpack-plugin==0.4
djangorestframework==3.9.0
flower==1.0.0
httplib2==0.9.2
idna==2.7
itypes==1.1.0
Jinja2==2.10
keyring==10.6.0
keyrings.alt==3.0
kombu==4.2.1
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
MarkupSafe==1.1.0
marshmallow==2.16.3
mysql-connector-python==8.0.13
oauth==1.0.1
oauthlib==2.1.0
olefile==0.45.1
optional-django==0.1.0
pbr==5.1.1
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.0.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
Pygments==2.3.0
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyOpenSSL==17.0.0
pyRFC3339==1.0
python-apt==1.6.3
python-dateutil==2.7.3
python-debian==0.1.32
python-editor==1.0.3
python3-openid==3.1.0
pytz==2018.7
pyxdg==0.25
PyYAML==3.12
react==4.3.0
redis==2.10.6
reportlab==3.4.0
requests==2.20.1
requests-oauthlib==1.0.0
requests-unixsocket==0.1.5
resumable-urlretrieve==0.1.5
SecretStorage==2.3.1
simplejson==3.13.2
six==1.11.0
SQLAlchemy==1.2.12
stevedore==1.30.0
system-service==0.3
systemd-python==234
tornado==4.5.3
ubuntu-drivers-common==0.0.0
unattended-upgrades==0.1
uritemplate==3.0.0
urllib3==1.24.1
usb-creator==0.3.3
vine==1.1.4
virtualenv==16.1.0
virtualenv-clone==0.4.0
virtualenvwrapper==4.8.2
wadllib==1.3.2
Werkzeug==0.14.1
xkit==0.0.0
zope.interface==4.3.2
